requirements.txt
streamlit
openai
llama-cloud==0.1.10
llama-index==0.12.12
llama-index-agent-openai==0.4.2
llama-index-cli==0.4.0
llama-index-core==0.12.14
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.4
llama-index-llms-openai==0.3.14
llama-index-multi-modal-llms-openai==0.4.2
llama-index-postprocessor-flag-embedding-reranker==0.3.0
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.3
llama-index-readers-llama-parse==0.4.0
llama-parse==0.5.20
llamaindex-py-client==0.1.19
nest-asyncio==1.6.0
openai==1.60.0
python-dotenv
requests==2.32.3
git+https://github.com/FlagOpen/FlagEmbedding.git
docx2txt
openpyxl
llama-index-embedding-gemini
llama-index-llms-gemini
llama-index-llms-deepseek